Remove use of typing.Final.
astruct had special-cased this to indicate class constants, but
that's not its intended meaning. Remove them throughout and change
astruct to understand that fields with values assigned at class
definition time with names in all caps are probably constants.

Also update the mypy plugin to allow non-Final ints to serve as the
parameters to astruct's sized ctypes aliases.